amd gpu memory deallocation
